+++ src/main/java/com/wandted/matitnyam/domain/entity/User.java
@@ -2,10 +2,10 @@ package com.wandted.matitnyam.domain.entity;
 
 
 import java.io.Serial;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EntityListeners;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -41,14 +41,14 @@ public class User extends AbstractEntity {
     @Column(name = "address", nullable = false)
     private String address;
     @Column(name = "longitude", nullable = false)
-    private Float longitude;
+    private Double longitude;
     @Column(name = "latitude", nullable = false)
-    private Float latitude;
+    private Double latitude;
     @Column(name = "suggest_alarm", nullable = false)
     private Boolean suggestAlarm;
 
     @Builder
-    public User(String id, String password, String address, Float longitude, Float latitude, Boolean suggestAlarm) {
+    public User(String id, String password, String address, Double longitude, Double latitude, Boolean suggestAlarm) {
         this.id = id;
         this.password = password;
         this.address = address;
